requirements.txt
numpy==1.23.5
streamlit==1.22.0
Pillow==9.5.0
keras==2.12.0
tensorflow==2.12.0
numpy==1.24.3
torchvision==0.15.1
opencv-python==4.6.0.66
matplotlib==3.5.3
plotly==5.15.0
git+https://github.com/facebookresearch/detectron2.git
